pyproject.toml
[tool.poetry]
package-mode = false
name = "house-of-wisdom"
version = "0.1.0"
description = "House Of Wisdom, an innovative Learning Management System (LMS)."
authors = ["Yousef Abu Shanab <josephyousef249@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "how" }]

[tool.poetry.dependencies]
python = ">=3.10"
django = ">=5.2.1"
django-cors-headers = ">=4.7.0"
django-filter = ">=25.1"
djangorestframework = ">=3.16.0"
markdown = ">=3.8"
pygments = ">=2.19.1"
djoser = ">=2.3.1"
drf-restwind = ">=0.1.8"
huggingface-hub = ">=0.31.1"
pillow = ">=11.2.1"
python-dotenv = ">=1.1.0"
wagtail = ">=7.0"
wagtailcodeblock = ">=1.29.0.2"
wagtail-localize = ">=1.12.1"

[tool.poetry.group.dev.dependencies]
black = ">=25.1.0"
ruff = ">=0.11.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
